Enforce shared tiers for urls
urls is the bottom of the shared-utils tier and had three imports from
above it: the generated-entity types from metabase/api, and DimensionType
plus encodeState from the metrics-viewer state module. The url builders
now declare the few generated-entity fields they read, and the function
that encodes a metrics-viewer page state into a URL moves into
common/metrics-viewer next to encodeState. That takes urls to zero
violations, so its enforceSharedTiers: false line is deleted.
